requirements.txt
# Core Dependencies
anthropic>=0.18.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Web & API
requests>=2.31.0
beautifulsoup4>=4.12.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Utilities
colorama>=0.4.6
rich>=13.7.0
